pytest: Separate forms tests
